backend/requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
djangorestframework-simplejwt==5.2.2
setuptools==70.0.0
django-cors-headers==4.7.0